+++ sdk/agentserver/azure-ai-agentserver-core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release History
 
+
+## 1.0.0b3 (2025-11-11)
+
+### Bugs Fixed
+
+- Fixed Id generator format.
+
+- Fixed trace initialization for agent-framework.
+
+
+## 1.0.0b2 (2025-11-10)
+
+### Bugs Fixed
+
+- Fixed Id generator format.
+
+- Improved stream mode error messsage.
+
+- Updated application insights related configuration environment variables.
+
+
 ## 1.0.0b1 (2025-11-07)
 
 ### Features Added